adding support for arm clang binaries

+++ scripts/pinned_build.sh
@@ -49,12 +49,17 @@ install_clang() {
    if [ ! -d "${CLANG_DIR}" ]; then
       echo "Installing Clang ${CLANG_VER} @ ${CLANG_DIR}"
       mkdir -p ${CLANG_DIR}
-      wget https://github.com/llvm/llvm-project/releases/download/llvmorg-${CLANG_VER}/clang+llvm-${CLANG_VER}-x86_64-linux-gnu-ubuntu-16.04.tar.xz
-      tar -xvf clang+llvm-${CLANG_VER}-x86_64-linux-gnu-ubuntu-16.04.tar.xz -C ${CLANG_DIR}
+      if [[ ${ARCH} = "x86_64" ]]; then
+         CLANG_FN=clang+llvm-${CLANG_VER}-x86_64-linux-gnu-ubuntu-16.04.tar.xz
+      else
+         CLANG_FN=clang+llvm-${CLANG_VER}-aarch64-linux-gnu.tar.xz
+      fi
+      wget https://github.com/llvm/llvm-project/releases/download/llvmorg-${CLANG_VER}/${CLANG_FN}
+      tar -xvf ${CLANG_FN} -C ${CLANG_DIR}
       pushdir ${CLANG_DIR}
       mv clang+*/* .
       popdir ${DEP_DIR}
-      rm clang+llvm-${CLANG_VER}-x86_64-linux-gnu-ubuntu-16.04.tar.xz
+      rm ${CLANG_FN}
    fi
    export PATH=${CLANG_DIR}/bin:$PATH
    export CLANG_DIR=${CLANG_DIR}
@@ -62,22 +67,24 @@ install_clang() {
 
 install_llvm() {
    LLVM_DIR=$1
-   if [ ! -d "${LLVM_DIR}" ]; then
-      echo "Installing LLVM ${LLVM_VER} @ ${LLVM_DIR}"
-      mkdir -p ${LLVM_DIR}
-      wget https://github.com/llvm/llvm-project/releases/download/llvmorg-${LLVM_VER}/llvm-${LLVM_VER}.src.tar.xz
-      tar -xvf llvm-${LLVM_VER}.src.tar.xz
-      pushdir "${LLVM_DIR}.src"
-      pushdir build
-      cmake -DCMAKE_TOOLCHAIN_FILE=${SCRIPT_DIR}/pinned_toolchain.cmake -DCMAKE_INSTALL_PREFIX=${LLVM_DIR} -DLLVM_TARGETS_TO_BUILD=host -DLLVM_BUILD_TOOLS=Off -DLLVM_ENABLE_RTTI=On -DLLVM_ENABLE_TERMINFO=Off -DCMAKE_EXE_LINKER_FLAGS=-pthread -DCMAKE_SHARED_LINKER_FLAGS=-pthread -DLLVM_ENABLE_PIC=NO ..
-      make -j${JOBS} 
-      make -j${JOBS} install
-      popdir "${LLVM_DIR}.src"
-      popdir ${DEP_DIR}
-      rm -rf ${LLVM_DIR}.src 
-      rm llvm-${LLVM_VER}.src.tar.xz
+   if [[ ${ARCH} = "x86_64" ]]; then
+      if [ ! -d "${LLVM_DIR}" ]; then
+         echo "Installing LLVM ${LLVM_VER} @ ${LLVM_DIR}"
+         mkdir -p ${LLVM_DIR}
+         wget https://github.com/llvm/llvm-project/releases/download/llvmorg-${LLVM_VER}/llvm-${LLVM_VER}.src.tar.xz
+         tar -xvf llvm-${LLVM_VER}.src.tar.xz
+         pushdir "${LLVM_DIR}.src"
+         pushdir build
+         cmake -DCMAKE_TOOLCHAIN_FILE=${SCRIPT_DIR}/pinned_toolchain.cmake -DCMAKE_INSTALL_PREFIX=${LLVM_DIR} -DLLVM_TARGETS_TO_BUILD=host -DLLVM_BUILD_TOOLS=Off -DLLVM_ENABLE_RTTI=On -DLLVM_ENABLE_TERMINFO=Off -DCMAKE_EXE_LINKER_FLAGS=-pthread -DCMAKE_SHARED_LINKER_FLAGS=-pthread -DLLVM_ENABLE_PIC=NO ..
+         make -j${JOBS} 
+         make -j${JOBS} install
+         popdir "${LLVM_DIR}.src"
+         popdir ${DEP_DIR}
+         rm -rf ${LLVM_DIR}.src 
+         rm llvm-${LLVM_VER}.src.tar.xz
+      fi
+      export LLVM_DIR=${LLVM_DIR}
    fi
-   export LLVM_DIR=${LLVM_DIR}
 }
 
 install_boost() {
@@ -103,12 +110,8 @@ install_boost() {
 
 pushdir ${DEP_DIR} # dir stack <DEP_DIR>/
 
-if [[ ${ARCH} = "x86_64" ]]; then
-   install_clang ${DEP_DIR}/clang-${CLANG_VER}
-   install_llvm ${DEP_DIR}/llvm-${LLVM_VER}
-   CMAKE_ARG="-DCMAKE_TOOLCHAIN_FILE=${SCRIPT_DIR}/pinned_toolchain.cmake"
-fi
-
+install_clang ${DEP_DIR}/clang-${CLANG_VER}
+install_llvm ${DEP_DIR}/llvm-${LLVM_VER}
 install_boost ${DEP_DIR}/boost_${BOOST_VER//\./_} "-with-toolset=clang"
 
 popdir ${SCRIPT_DIR}
@@ -119,6 +122,10 @@ MANDEL_DIR=${DEP_DIR}/mandel
 
 echo "Building Mandel"
 pushdir ../build
+
+if [[ ${ARCH} = "x86_64" ]]; then
+   CMAKE_ARG="-DCMAKE_TOOLCHAIN_FILE=${SCRIPT_DIR}/pinned_toolchain.cmake"
+fi
 
 cmake ${CMAKE_ARG} -DCMAKE_INSTALL_PREFIX=/usr/local -DCMAKE_BUILD_TYPE=Release -DCMAKE_PREFIX_PATH=${LLVM_DIR}/lib/cmake -DCMAKE_PREFIX_PATH=${BOOST_DIR}/bin ..
